requirements.txt
spark
python3-distutils
python3-apt
pandas
google-cloud
google-cloud-bigquery
mage-ai


